Move context in store folder
