+++ src/index-server.tsx
@@ -4,7 +4,7 @@ import { routes } from "./routes"
 import App from "./components/app/App"
 import { langServiceInstance } from "./LangService"
 import { requestStaticPropsFromRoute, Router } from "@cher-ami/router"
-import { GlobalDataContext } from "./GlobalDataContext"
+import { GlobalDataContext } from "./store/GlobalDataContext"
 import { preventSlashes } from "../config/helpers/prevent-slashes.js"
 import { loadEnv } from "vite"
 import { TScriptsObj } from "../prerender/helpers/ManifestParser"
